+++ CHANGELOG.md
@@ -9,10 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `EnumerableSet` trait. #687
+- Add macro to generate `EnumerableSet` for: `Address`, `B256`, `U8`, `U16`, `U32`, `U64`, `U128`, `U256` . #687
 - Add `IErc1155Receiver` trait. #747
 - Add `Erc1155Holder` contract. #747
 - Add `IErc721Receiver` trait. #743
 - Add `Erc721Holder` contract. #743
+
+### Changed
+
+- Remove initial `EnumerableAddressSet` implementation. #687
 
 ### Changed (Breaking)